Refactor wxAccessible code as it handles more than sliders.
