+++ src/CMakeLists.txt
@@ -1,5 +1,4 @@
 set(FREEDV_SOURCES
-    DecibelSliderAccessible.cpp
     dlg_audiooptions.cpp
     dlg_easy_setup.cpp
     dlg_filter.cpp
@@ -40,8 +39,6 @@ set(FREEDV_SOURCES
     ongui.cpp
     freedv_interface.cpp
     freedv_interface.h        
-    wxListViewComboPopup.cpp
-    wxListViewComboPopup.h
 )
 
 set(FREEDV_SOURCES_OSX
@@ -55,6 +52,7 @@ set(FREEDV_LINK_LIBS_OSX
 # Compile FreeDV components
 add_subdirectory(audio)
 add_subdirectory(config)
+add_subdirectory(gui)
 add_subdirectory(pipeline)
 add_subdirectory(reporting)
 add_subdirectory(sox)
@@ -82,9 +80,9 @@ endif(APPLE)
 
 # Link imported or build tree targets.
 if(APPLE)
-target_link_libraries(FreeDV fdv_audio fdv_audio_pipeline fdv_config fdv_reporting fdv_sox lpcnetfreedv codec2)
+target_link_libraries(FreeDV fdv_audio fdv_audio_pipeline fdv_config fdv_gui_util fdv_reporting fdv_sox lpcnetfreedv codec2)
 else(APPLE)
-target_link_libraries(freedv fdv_audio fdv_audio_pipeline fdv_config fdv_reporting fdv_sox lpcnetfreedv codec2)
+target_link_libraries(freedv fdv_audio fdv_audio_pipeline fdv_config fdv_gui_util fdv_reporting fdv_sox lpcnetfreedv codec2)
 endif(APPLE)
 
 # Add build dependencies for internally built external libraries.
